requirements.txt
# Core Flask dashboard
Flask==3.1.0
Flask-HTTPAuth
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0

# Gmail API + OAuth2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
proto-plus==1.26.1
protobuf==6.30.2
httplib2==0.22.0

# AI clients
openai==1.74.0
requests==2.32.3
urllib3==2.4.0

# .env support
python-dotenv==1.1.0

# CSV and JSON parsing support
pandas==2.2.3
numpy==2.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# HTTP + OAuth support
requests-oauthlib==2.0.0
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2

# ASGI / HTTP handling (optional Firework or FastAPI integration)
httpx==0.28.1
httpcore==1.0.8
h11==0.14.0
anyio==4.9.0
sniffio==1.3.1

# Misc / CLI / Dev
tqdm==4.67.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
typing_extensions==4.13.2
typing-inspection==0.4.0
annotated-types==0.7.0
blinker==1.9.0
idna==3.10
charset-normalizer==3.4.1
cachetools==5.5.2
jiter==0.9.0
